refactoring WSRefInjectionSource
